+++ daiv/daiv/__init__.py
@@ -2,6 +2,6 @@
 # Django starts so that shared_task will use this app.
 from .celeryapp import app as celery_app
 
-__version__ = "0.1.0-alpha.11"
+__version__ = "0.1.0-alpha.12"
 
 __all__ = ("celery_app",)